# pkg: improve get forwarded host